requirement.txt
streamlit
yfinance
pandas
numpy
langchain_pinecone
openai
python-dotenv 
langchain-community 
sentence_transformers
request